Improved upon previous design
createApp(GameData, CustomAppliance) creates an Appliance using values provided from CustomAppliance.
If BaseApplianceID is not set, it will default to a counter.
This compares the values in CustomAppliance and an Empty Appliance, and if there are any differences they are applied to the final result.

This allows for more customisation of the CustomAppliances. Previously we could only apply some basic changes such as Name, Description, etc. We can now change every field available.